backend/requirements.txt
# app
Django
psycopg2-binary
celery
django-redis
requests
django-debug-toolbar
django-environ

# utility
setuptools-scm

# opentelemetry
opentelemetry-api
opentelemetry-sdk
opentelemetry-distro[oltp]

# instrumentation
opentelemetry-instrumentation
opentelemetry-instrumentation-django
opentelemetry-instrumentation-celery
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-urllib
opentelemetry-instrumentation-urllib3

# exporter
opentelemetry-exporter-otlp-proto-grpc
protobuf~=3.20.0  # opentelemetry-exporter-otlp-proto-grpc requires 3.20.0 or lower
